Split post 1428 oversized slices
