+++ app/src/main/java/com/pocketshell/app/tmux/TmuxSessionInputHandlers.kt
@@ -1,0 +1,93 @@
+package com.pocketshell.app.tmux
+
+import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.gestures.detectVerticalDragGestures
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.unit.dp
+import kotlin.math.abs
+
+internal val VerticalSwipeThreshold = 72.dp
+
+internal fun Modifier.verticalSwipeInput(
+    thresholdPx: Float,
+    onBoundary: () -> Unit,
+    onSwipeUp: (() -> Unit)? = null,
+    onSwipeDown: (() -> Unit)? = null,
+) = pointerInput(thresholdPx, onBoundary, onSwipeUp, onSwipeDown) {
+    var totalDrag = 0f
+    var triggered = false
+    detectVerticalDragGestures(
+        onDragStart = {
+            totalDrag = 0f
+            triggered = false
+        },
+        onVerticalDrag = { change, dragAmount ->
+            totalDrag += dragAmount
+            if (triggered || abs(totalDrag) < thresholdPx) {
+                return@detectVerticalDragGestures
+            }
+            val swipeUp = totalDrag < 0f
+            val callback = if (swipeUp) onSwipeUp else onSwipeDown
+            if (callback == null) return@detectVerticalDragGestures
+            triggered = true
+            onBoundary()
+            callback()
+            change.consume()
+        },
+    )
+}
+
+/**
+ * Issue #167: a thin wrapper around [BackHandler] that codifies the back
+ * routing on [TmuxSessionScreen].
+ *
+ * Order of precedence:
+ *  1. If a lifecycle dialog is open ([dialogOpen]), back closes the
+ *     dialog (mirroring the Cancel button).
+ *  2. Otherwise, if the session drawer is open ([sessionDrawerOpen]),
+ *     back closes the drawer.
+ *  3. Otherwise, the window switcher / mic sheet / snippet picker
+ *     overlays close in turn.
+ *  4. Otherwise [onBack] runs - which pops the in-app back stack to the
+ *     host list. The hosted `TmuxSessionViewModel` is then cleared by
+ *     Compose's lifecycle owner, and its `onCleared()` tears the SSH +
+ *     tmux client state down (via `closeCurrentConnection`).
+ *
+ * Extracted as its own composable so the routing behaviour can be
+ * regression-tested without a Hilt graph or a live tmux client - the
+ * AVD-side reviewer evidence is the user journey "attach -> system back ->
+ * host list" against the deterministic Docker fixture, which exercises
+ * the implicit teardown via the ViewModel's `onCleared`.
+ *
+ * `DropdownMenu` and `AlertDialog` already intercept system-back
+ * themselves in front of any `BackHandler` registered by the screen, so
+ * the more-menu and window-context-menu dropdowns are not surfaced here;
+ * back on those closes the popup first by the dropdown's own handling.
+ */
+@Composable
+internal fun TmuxSessionBackHandler(
+    dialogOpen: Boolean,
+    sessionDrawerOpen: Boolean,
+    micSheetOpen: Boolean,
+    snippetPickerOpen: Boolean,
+    cardFeedSheetOpen: Boolean = false,
+    onDismissDialog: () -> Unit,
+    onDismissSessionDrawer: () -> Unit,
+    onDismissMicSheet: () -> Unit,
+    onDismissSnippetPicker: () -> Unit,
+    onDismissCardFeedSheet: () -> Unit = {},
+    onBack: () -> Unit,
+) {
+    BackHandler {
+        when {
+            dialogOpen -> onDismissDialog()
+            sessionDrawerOpen -> onDismissSessionDrawer()
+            micSheetOpen -> onDismissMicSheet()
+            snippetPickerOpen -> onDismissSnippetPicker()
+            cardFeedSheetOpen -> onDismissCardFeedSheet()
+            else -> onBack()
+        }
+    }
+}
