+++ app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
@@ -226,417 +226,6 @@ class TmuxSessionViewModelTest : TmuxSessionViewModelTestBase() {
     // tests are deleted with the feature. The conversation feed now shows every
     // event with no query filter. ─────────────
 
-    // ─── Issue #186: per-window agent detection state ──────────────────
-    //
-    // Detection is per-pane (and therefore per-window for the simple
-    // case of one pane per window). The view-model surface that the
-    // screen drives off is [agentForWindow] — it must return the
-    // current window's agent kind regardless of which window's pane the
-    // user is currently viewing, so the Conversation tab can hide on
-    // plain-shell windows even when a sibling window has a live agent.
-
-    @Test
-    fun agentForWindowReturnsNullForUnknownWindowId() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-
-        assertNull(vm.agentForWindow(null))
-        assertNull(vm.agentForWindow(""))
-        assertNull(vm.agentForWindow("@nonexistent"))
-    }
-
-    @Test
-    fun agentForWindowReturnsKindOfDetectedAgentInThatWindow() = runTest(scheduler) {
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%0",
-                    windowId = "@0",
-                    sessionId = "$0",
-                    title = "agent",
-                    paneIndex = 0,
-                ),
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%1",
-                    windowId = "@1",
-                    sessionId = "$0",
-                    title = "shell",
-                    paneIndex = 0,
-                ),
-            ),
-        )
-        // Only window @0's pane has a detection — window @1 is a plain
-        // shell with no agent.
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-
-        assertEquals(
-            "the agent-running window must report its agent kind",
-            AgentKind.ClaudeCode,
-            vm.agentForWindow("@0"),
-        )
-        assertNull(
-            "a plain-shell window must NOT inherit a sibling window's agent kind " +
-                "even when they share a cwd",
-            vm.agentForWindow("@1"),
-        )
-    }
-
-    @Test
-    fun agentForWindowPicksLowestPaneIndexWhenMultipleAgentsInOneWindow() = runTest(scheduler) {
-        // Edge case: two panes in the same window with detections.
-        // Stable behaviour: the pane that appears first in the panes
-        // list wins (which is paneIndex ascending per the reconcile
-        // sort).
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%0",
-                    windowId = "@0",
-                    sessionId = "$0",
-                    title = "a",
-                    paneIndex = 0,
-                ),
-                TmuxSessionViewModel.ParsedPane(
-                    paneId = "%1",
-                    windowId = "@0",
-                    sessionId = "$0",
-                    title = "b",
-                    paneIndex = 1,
-                ),
-            ),
-        )
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        vm.startAgentConversationForTest(
-            "%1",
-            AgentDetection(
-                agent = AgentKind.Codex,
-                sourcePath = "/home/u/.codex/sessions/x.jsonl",
-                sessionId = "x",
-                confidence = AgentDetection.Confidence.ProcessConfirmed,
-            ),
-        )
-
-        assertEquals(AgentKind.ClaudeCode, vm.agentForWindow("@0"))
-    }
-
-    @Test
-    fun listPanesParserAcceptsPrintableFieldSeparator() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        client.responses.addLast(
-            CommandResponse(
-                number = 1L,
-                output = listOf(
-                    listOf(
-                        "%0",
-                        "@0",
-                        "\$0",
-                        "work",
-                        "shell",
-                        "0",
-                        "/workspace",
-                        "bash",
-                        "/dev/pts/3",
-                    ).joinToString(LIST_PANES_FIELD_SEPARATOR),
-                ),
-                isError = false,
-            ),
-        )
-        vm.replaceClientForTest(
-            hostId = 1L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        client.emittedEvents.emit(
-            ControlEvent.WindowAdd(sessionId = "\$0", windowId = "@0", name = ""),
-        )
-        advanceUntilIdle()
-
-        val pane = vm.panes.value.single()
-        assertEquals("%0", pane.paneId)
-        assertEquals("/workspace", pane.cwd)
-        assertEquals("bash", pane.currentCommand)
-        assertEquals("/dev/pts/3", pane.paneTty)
-    }
-
-    @Test
-    fun listPanesFormatRequestsPaneTty() = runTest(scheduler) {
-        // Issue #186: the list-panes format must include
-        // `#{pane_tty}` so per-pane detection can scope its process
-        // scan to a TTY. Without this, the per-window fix has no
-        // signal to act on.
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        client.responses.addLast(
-            CommandResponse(
-                number = 1L,
-                output = listOf("%0\t@0\t\$0\twork\tshell\t0\t/workspace\tbash\t/dev/pts/3"),
-                isError = false,
-            ),
-        )
-        vm.replaceClientForTest(
-            hostId = 1L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        client.emittedEvents.emit(
-            ControlEvent.WindowAdd(sessionId = "\$0", windowId = "@0", name = ""),
-        )
-        advanceUntilIdle()
-
-        val listPanesCmd = client.sentCommands.single { it.startsWith("list-panes") }
-        assertTrue(
-            "list-panes format must include #{pane_tty}; got `$listPanesCmd`",
-            listPanesCmd.contains("#{pane_tty}"),
-        )
-        assertTrue(
-            "list-panes format must include #{pane_in_mode}; got `$listPanesCmd`",
-            listPanesCmd.contains("#{pane_in_mode}"),
-        )
-
-        val pane = vm.panes.value.single()
-        assertEquals("/dev/pts/3", pane.paneTty)
-        assertFalse(pane.inCopyMode)
-    }
-
-    @Test
-    fun listPanesParserCapturesTmuxWindowIndex() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        client.responses.addLast(
-            CommandResponse(
-                number = 1L,
-                output = listOf("%0\t@7\t2\t\$0\twork\tshell\t0\t/workspace\tbash\t/dev/pts/3\t0"),
-                isError = false,
-            ),
-        )
-        vm.replaceClientForTest(
-            hostId = 1L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-        client.emittedEvents.emit(
-            ControlEvent.WindowAdd(sessionId = "\$0", windowId = "@7", name = ""),
-        )
-        advanceUntilIdle()
-
-        val pane = vm.panes.value.single()
-        assertEquals("@7", pane.windowId)
-        assertEquals(2, pane.windowIndex)
-    }
-
-    @Test
-    fun paneModeChangedRefreshesCopyModeStateFromListPanes() = runTest(scheduler) {
-        val vm = newVm()
-        val client = FakeTmuxClient()
-        client.responses.addLast(
-            CommandResponse(
-                number = 1L,
-                output = listOf("%0\t@0\t\$0\twork\tshell\t0\t/workspace\tbash\t/dev/pts/3\t1"),
-                isError = false,
-            ),
-        )
-        vm.replaceClientForTest(
-            hostId = 1L,
-            hostName = "alpha",
-            host = "alpha.example",
-            port = 22,
-            user = "alex",
-            keyPath = "/keys/a",
-            sessionName = "work",
-            client = client,
-        )
-
-        client.emittedEvents.emit(ControlEvent.PaneModeChanged("%0"))
-        advanceUntilIdle()
-
-        val pane = vm.panes.value.single()
-        assertEquals("%0", pane.paneId)
-        assertTrue("pane_in_mode=1 must mark the pane as copy-mode/action mode", pane.inCopyMode)
-    }
-
-    @Test
-    fun agentForWindowClearsAfterPaneLosesDetection() = runTest(scheduler) {
-        // The screen drives Conversation-tab visibility off
-        // [agentForWindow]. When a pane that previously had a
-        // detection no longer reports one (the user exited Claude),
-        // `agentForWindow` for that window must return null so the tab
-        // disappears.
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "a", paneIndex = 0),
-            ),
-        )
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        assertEquals(AgentKind.ClaudeCode, vm.agentForWindow("@0"))
-
-        // Simulate the production clear path: drop the conversation
-        // entry directly (mirrors what `clearAgentDetectionForPane`
-        // does when a subsequent source-resolution probe returns null).
-        vm.clearAgentDetectionForPaneForTest("%0")
-
-        assertNull(
-            "agentForWindow must return null once the pane's detection is cleared",
-            vm.agentForWindow("@0"),
-        )
-    }
-
-    @Test
-    fun conversationStateClearsWhenPaneLosesDetection() = runTest(scheduler) {
-        // If the user has opened the Conversation tab on the agent pane
-        // and then exits the agent in that pane, the per-pane
-        // conversation state must disappear so the screen falls back to
-        // Terminal for that visible pane.
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "agent", paneIndex = 0),
-            ),
-        )
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        assertEquals(SessionTab.Conversation, vm.agentConversations.value["%0"]!!.selectedTab)
-
-        vm.clearAgentDetectionForPaneForTest("%0")
-
-        assertNull(
-            "conversation state must clear when the pane loses its detection",
-            vm.agentConversations.value["%0"],
-        )
-        assertNull(
-            "agentForWindow must report null once the pane lost its detection",
-            vm.agentForWindow("@0"),
-        )
-    }
-
-    @Test
-    fun conversationWithLoadedTranscriptIsKeptReadableWhenDetectionDrops() = runTest(scheduler) {
-        // Issue #1057 (maintainer dogfood — "conversation is not visible in this
-        // app"): a conversation that genuinely EXISTS (a loaded transcript, events
-        // present) must stay READABLE after live detection drops, so the
-        // Terminal/Conversation toggle stays reachable and the user can still read
-        // it. On base clearAgentDetectionForPane DROPPED the row the instant
-        // detection settled null → the transcript became unreachable (the exact
-        // symptom). This is the fast JVM red→green for the VM fix; the connected
-        // sibling is ConversationStaysReachableAfterDetectionDropsDockerTest.
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "agent", paneIndex = 0),
-            ),
-        )
-        val event = ConversationEvent.Message(
-            id = "m1",
-            agent = AgentKind.ClaudeCode,
-            role = ConversationRole.Assistant,
-            text = "the transcript the user was reading",
-        )
-        vm.startAgentConversationForTest("%0", newClaudeDetection(), listOf(event))
-        vm.selectSessionTab("%0", SessionTab.Conversation)
-        runCurrent()
-        assertEquals(
-            SessionTab.Conversation,
-            vm.agentConversations.value["%0"]!!.selectedTab,
-        )
-
-        // Live detection settles null (the agent exited / re-detection never
-        // rebinds) — the production teardown the null-detection poll calls.
-        vm.clearAgentDetectionForPaneForTest("%0")
-        runCurrent()
-
-        val row = vm.agentConversations.value["%0"]
-        assertNotNull(
-            "#1057: an events-bearing conversation row is KEPT readable after " +
-                "detection drops (on base it was dropped → conversation unreachable)",
-            row,
-        )
-        assertNull(
-            "#1057: the kept row's detection is null (the live agent is gone)",
-            row!!.detection,
-        )
-        assertTrue(
-            "#1057: the kept row preserves the loaded transcript so it stays readable",
-            row.events.isNotEmpty(),
-        )
-        assertEquals(
-            "#1057: the user's Conversation tab choice persists on the kept row",
-            SessionTab.Conversation,
-            row.selectedTab,
-        )
-        // agentForWindow still reports null (no LIVE agent) — the kept row is a
-        // frozen transcript, not a resurrected live detection.
-        assertNull(
-            "#1057: a kept frozen transcript must not resurrect a live agent",
-            vm.agentForWindow("@0"),
-        )
-    }
-
-    @Test
-    fun conversationWithNoTranscriptStillDropsWhenDetectionDrops() = runTest(scheduler) {
-        // Issue #1057 adjacency (#186/#894 contract unchanged): a row with NO
-        // loaded transcript (the agent exited before any transcript was read, a
-        // genuine shell) still DROPS when detection settles null — the keep-events
-        // change must NOT make a conversation-less window linger on the toggle.
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(
-                TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "agent", paneIndex = 0),
-            ),
-        )
-        vm.startAgentConversationForTest("%0", newClaudeDetection())
-        runCurrent()
-        assertNotNull(vm.agentConversations.value["%0"])
-
-        vm.clearAgentDetectionForPaneForTest("%0")
-        runCurrent()
-
-        assertNull(
-            "#1057 adjacency: a row with no loaded transcript still drops on exit",
-            vm.agentConversations.value["%0"],
-        )
-        assertNull(vm.agentForWindow("@0"))
-    }
-
-    @Test
-    fun parsedPanePaneTtyDefaultsToEmptyWhenOmitted() = runTest(scheduler) {
-        // Defensive: an older tmux that doesn't emit the new field, or
-        // a unit test passing the legacy shape, must still produce a
-        // valid TmuxPaneState (with an empty paneTty so per-pane
-        // detection short-circuits to null).
-        val vm = newVm()
-        vm.attachClientForTest(FakeTmuxClient())
-        vm.applyParsedPanesForTest(
-            listOf(TmuxSessionViewModel.ParsedPane("%0", "@0", "$0", "a", paneIndex = 0)),
-        )
-        advanceUntilIdle()
-
-        assertEquals("", vm.panes.value.single().paneTty)
-    }
-
     // ─── Issue #178: same-host fast-switch reuses the SSH transport ───
     //
     // The connect() path detects "same host, different tmux session"
